infosys project/-Building-an-Ambient-Agent-with-LangGraph-for-an-Email-Assistant/requirements.txt
langgraph
langchain
langchain-google-genai
langsmith
python-dotenv
pydantic